incremento de código

+++ coletadelixo/src/main/java/br/computacao/coletadelixo/dao/ColetaDao.java
@@ -2,6 +2,6 @@ package br.computacao.coletadelixo.dao;
 
 import br.computacao.coletadelixo.model.Coleta;
 
-public class ColetaDao extends Dao<Coleta,Long>{
+public class ColetaDao extends Dao<Coleta,Integer>{
 
 }
