+++ coletadelixo/src/main/java/br/computacao/coletadelixo/dao/EmpresaDao.java
@@ -2,6 +2,6 @@ package br.computacao.coletadelixo.dao;
 
 import br.computacao.coletadelixo.model.Empresa;
 
-public class EmpresaDao extends Dao<Empresa,Long> { 
+public class EmpresaDao extends Dao<Empresa,Integer> { 
 
 }
