+++ coletadelixo/src/main/java/br/computacao/coletadelixo/dao/MoradorDao.java
@@ -2,6 +2,6 @@ package br.computacao.coletadelixo.dao;
 
 import br.computacao.coletadelixo.model.Morador;
 
-public class MoradorDao extends Dao <Morador, Long>{
+public class MoradorDao extends Dao <Morador, Integer>{
 
 }
